\name{QCRep}
\alias{QCRep}
\title{Illuminas Infinium HumanMethylation27 Internal Controls}
\description{
Creates all the histogram plots relative to the internal controls of the Illumina assay into pdf files.
}
\usage{
QCRep(Dir)
}
\arguments{
  \item{Dir}{A character string containing the location of the directory 
    in which the input file are
}
}

\details{
 This creates all the histograms foreach internal control. The sample-independent controls let you evaluate the quality of specific
steps in the process flow; instead, the sample-dependent controls let you evaluate performance across samples. }

\author{ F.M. Mancuso <francesco.mancuso@crg.es> }

\keyword{ hplot }
\keyword{ manip }
